Move StockController into Admin/Sell/Catalog/
