+++ src/PrestaShopBundle/Resources/views/Admin/Sell/Catalog/Stock/overview.html.twig
